requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
filelock==3.12.2
Flask==2.2.5
Flask-Cors==4.0.0
Flask-JSON==0.4.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.4.0
gunicorn==20.1.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
multidict==6.0.4
mysqlclient==2.2.0
openai==0.27.8
packaging==23.1
platformdirs==3.8.0
pluggy==1.2.0
PyJWT==2.8.0
pymongo==4.4.0
pytest==7.4.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
requests==2.31.0
simple-websocket==0.10.1
six==1.16.0
SQLAlchemy==2.0.19
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.23.1
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2